+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/adsb-setup/utils/agg_status.py
@@ -7,7 +7,26 @@ from enum import Enum
 from .util import generic_get_json, print_err, make_int
 from .data import Data
 
-T = Enum("T", ["Yes", "No", "Unknown"])
+T = Enum("T", ["Disconnected", "Unknown", "Good", "Bad", "Warning"])
+status_symbol = {
+    T.Disconnected: "\u2612",
+    T.Unknown: ".",
+    T.Good: "+",
+    T.Bad: "\u2639",
+    T.Warning: "\u26A0",
+}
+ultrafeeder_aggs = [
+    "adsblol",
+    "flyitaly",
+    "avdelphi",
+    "planespotters",
+    "tat",
+    "radarplane",
+    "adsbfi",
+    "adsbx",
+    "hpradar",
+    "alive",
+]
 
 
 class AggStatus:
@@ -30,7 +49,7 @@ class AggStatus:
         if not self.use_cached(now):
             self.check()
         if self.use_cached(now):
-            return "+" if self._beast == T.Yes else "-" if self._beast == T.No else "."
+            return status_symbol.get(self._beast, ".")
         return "."
 
     @property
@@ -39,7 +58,7 @@ class AggStatus:
         if not self.use_cached(now):
             self.check()
         if self.use_cached(now):
-            return "+" if self._mlat == T.Yes else "-" if self._mlat == T.No else "."
+            return status_symbol.get(self._mlat, ".")
         return "."
 
     def get_json(self, json_url):
@@ -77,8 +96,83 @@ class AggStatus:
             return response.text, response.status_code
         return None, status
 
+    def uf_path(self):
+        uf_dir = "/run/adsb-feeder-"
+        uf_dir += f"uf_{self._idx}" if self._idx != 0 else "ultrafeeder"
+        return uf_dir
+
+    def get_mlat_status(self):
+        if self._agg not in ultrafeeder_aggs:
+            self._mlat = T.Unknown
+            return
+        mconf = None
+        netconfig = self._d.netconfigs.get(self._agg)
+        if netconfig:
+            # example mlat config: "mlat,dati.flyitalyadsb.com,30100,39002",
+            mconf = netconfig.mlat_config
+        if not mconf:
+            self._mlat = T.Unknown
+            return
+        filename = f"{mconf.split(',')[1]}:{mconf.split(',')[2]}.json"
+        try:
+            mlat_json = json.load(open(f"{self.uf_path()}/mlat-client/{filename}", "r"))
+            percent_good = mlat_json.get("good_sync_percentage_last_hour", 0)
+            percent_bad = mlat_json.get("bad_sync_percentage_last_hour", 0)
+            peer_count = mlat_json.get("peer_count", 0)
+            now = mlat_json.get("now")
+        except:
+            print_err(f"checking {self.uf_path()}/mlat-client/{filename} failed")
+            self._mlat = T.Unknown
+            return
+        if now - int(datetime.now().timestamp()) > 60:
+            # that's more than a minute old... let's go with dunno
+            self._mlat = T.Unknown
+        elif peer_count == 0:
+            self._mlat = T.Disconnected
+        elif percent_good > 10 and percent_bad <= 2:
+            self._mlat = T.Good
+        elif percent_good == 0 or (percent_bad > 2 and percent_bad <= 10):
+            self._mlat = T.Warning
+        else:
+            self._mlat = T.Bad
+
+    def get_beast_status(self):
+        self._beast = T.Unknown
+        if self._agg not in ultrafeeder_aggs:
+            return
+        bconf = None
+        netconfig = self._d.netconfigs.get(self._agg)
+        if netconfig:
+            # example adsb config: "adsb,dati.flyitalyadsb.com,4905,beast_reduce_plus_out",
+            bconf = netconfig.adsb_config
+        if not bconf:
+            self._beast = T.Unknown
+            return
+        pattern = (
+            f"readsb_net_connector_status{{host=\"{bconf.split(',')[1]}\",port=\"{bconf.split(',')[2]}\"}} (\\d+)"
+        )
+        print(f"checking beast for {pattern}")
+        try:
+            readsb_status = json.reads(open(f"{self.uf_path()}/readsb/stats.prom", "r").read())
+        except:
+            self._beast = T.Unknown
+            return
+        match = re.search(pattern, readsb_status)
+        if match:
+            status = int(match.group(1))
+            if status == "-1":
+                self._beast = T.Disconnected
+            elif status >= "0":
+                self._beast = T.Good
+
     def check(self):
-        # figure out the feeder state at this aggregator (if possible)
+        # look up readsb / mlat_client view of the status
+        self.get_mlat_status()
+        self.get_beast_status()
+        # override this if we do get api results from that aggregator
+        # for beast status we prefer the aggregator results
+        # for mlat status we prefer the api results
+        api_mlat = T.Unknown
         if self._agg == "adsblol":
             uuid = self._d.env_by_tags("adsblol_uuid").list_get(self._idx)
             name = self._d.env_by_tags("site_name").list_get(self._idx)
@@ -89,21 +183,20 @@ class AggStatus:
                 if lolclients:
                     lolbeast = lolclients.get("beast")
                     lolmlat = lolclients.get("mlat")
-                    self._beast = T.No
+                    self._beast = T.Disconnected
                     if isinstance(lolbeast, list):
                         for entry in lolbeast:
                             if entry.get("uuid", "xxxxxxxx-xxxx-")[:14] == uuid[:14]:
-                                self._beast = T.Yes
+                                self._beast = T.Good
                                 self._d.env_by_tags("adsblol_link").list_set(self._idx, entry.get("adsblol_my_url"))
                                 break
-                    self._mlat = (
-                        T.Yes
+                    api_mlat = (
+                        T.Good
                         if isinstance(lolmlat, list)
                         and any(b.get("uuid", "xxxxxxxx-xxxx-")[:14] == uuid[:14] for b in lolmlat)
-                        else T.No
+                        else T.Disconnected
                     )
                     self._last_check = datetime.now()
-
                 else:
                     print_err(f"adsblol returned status {status}")
         if self._agg == "flyitaly":
@@ -113,8 +206,8 @@ class AggStatus:
             if response_dict and status == 200:
                 feeding = response_dict["feeding"]
                 if feeding:
-                    self._beast = T.Yes if feeding.get("beast") else T.No
-                    self._mlat = T.Yes if feeding.get("mlat") else T.No
+                    self._beast = T.Good if feeding.get("beast") else T.Disconnected
+                    api_mlat = T.Good if feeding.get("mlat") else T.Disconnected
                     self._last_check = datetime.now()
             else:
                 print_err(f"flyitaly returned {status}")
@@ -131,16 +224,16 @@ class AggStatus:
             adsbfi_dict, status = self.get_json(json_ip_url)
             if adsbfi_dict and status == 200:
                 mlat_array = adsbfi_dict.get("mlat", [])
-                self._mlat = T.Yes if any(m.get("user", "") == name for m in mlat_array) else T.No
+                api_mlat = T.Good if any(m.get("user", "") == name for m in mlat_array) else T.Disconnected
                 self._last_check = datetime.now()
             else:
                 print_err(f"adsbfi v1/myip returned {status}")
             adsbfi_dict, status = self.get_json(json_uuid_url)
             if adsbfi_dict and status == 200:
                 self._beast = (
-                    T.Yes
+                    T.Good
                     if len(adsbfi_dict.get("beast", [])) > 0 and adsbfi_dict.get("beast")[0].get("receiverId") == uuid
-                    else T.No
+                    else T.Disconnected
                 )
                 self._last_check = datetime.now()
             else:
@@ -151,8 +244,8 @@ class AggStatus:
             if radarplane_dict and status == 200:
                 rdata = radarplane_dict.get("data")
                 if rdata:
-                    self._beast = T.Yes if rdata.get("beast") else T.No
-                    self._mlat = T.Yes if rdata.get("mlat") else T.No
+                    self._beast = T.Good if rdata.get("beast") else T.Disconnected
+                    api_mlat = T.Good if rdata.get("mlat") else T.Disconnected
                     self._last_check = datetime.now()
             else:
                 print_err(f"radarplane returned {status}")
@@ -162,8 +255,14 @@ class AggStatus:
             fa_dict, status = self.get_json(json_url)
             if fa_dict and status == 200:
                 # print_err(f"fa status.json returned {fa_dict}")
-                self._beast = T.Yes if fa_dict.get("adept") and fa_dict.get("adept").get("status") == "green" else T.No
-                self._mlat = T.Yes if fa_dict.get("mlat") and fa_dict.get("mlat").get("status") == "green" else T.No
+                self._beast = (
+                    T.Good
+                    if fa_dict.get("adept") and fa_dict.get("adept").get("status") == "green"
+                    else T.Disconnected
+                )
+                api_mlat = (
+                    T.Good if fa_dict.get("mlat") and fa_dict.get("mlat").get("status") == "green" else T.Disconnected
+                )
                 self._last_check = datetime.now()
             else:
                 print_err(f"flightaware at {json_url} returned {status}")
@@ -173,7 +272,7 @@ class AggStatus:
             fr_dict, status = self.get_json(json_url)
             if fr_dict and status == 200:
                 # print_err(f"fr monitor.json returned {fr_dict}")
-                self._beast = T.Yes if fr_dict.get("feed_status") == "connected" else T.No
+                self._beast = T.Good if fr_dict.get("feed_status") == "connected" else T.Disconnected
                 self._last_check = datetime.now()
             else:
                 print_err(f"flightradar at {json_url} returned {status}")
@@ -182,8 +281,8 @@ class AggStatus:
             json_url = f"https://radarplane.com/api/v1/feed/check/{uuid}"
             rp_dict, status = self.get_json(json_url)
             if rp_dict and rp_dict.get("data") and status == 200:
-                self._beast = T.Yes if rp_dict["data"].get("beast") else T.No
-                self._mlat = T.Yes if rp_dict["data"].get("mlat") else T.No
+                self._beast = T.Good if rp_dict["data"].get("beast") else T.Disconnected
+                api_mlat = T.Good if rp_dict["data"].get("mlat") else T.Disconnected
                 self._last_check = datetime.now()
             else:
                 print_err(f"radarplane returned {status}")
@@ -218,8 +317,8 @@ class AggStatus:
                     if station:
                         online = station.get("online")
                         mlat_online = station.get("mlat_online")
-                        self._beast = T.Yes if online else T.No
-                        self._mlat = T.Yes if mlat_online else T.No
+                        self._beast = T.Good if online else T.Disconnected
+                        api_mlat = T.Good if mlat_online else T.Disconnected
                         self._last_check = datetime.now()
         elif self._agg == "1090uk":
             key = self._d.env_by_tags(["1090uk", "key"]).list_get(self._idx)
@@ -227,7 +326,7 @@ class AggStatus:
             tn_dict, status = self.get_json(json_url)
             if tn_dict and status == 200:
                 online = tn_dict.get("online", False)
-                self._beast = T.Yes if online else T.No
+                self._beast = T.Good if online else T.Disconnected
                 self._last_check = datetime.now()
         elif self._agg == "alive":
             json_url = "https://api.airplanes.live/feed-status"
@@ -237,18 +336,18 @@ class AggStatus:
                 beast_clients = a_dict.get("beast_clients")
                 # print_err(f"alife returned {beast_clients}", level=8)
                 if beast_clients:
-                    self._beast = T.Yes if any(bc.get("uuid") == uuid for bc in beast_clients) else T.No
+                    self._beast = T.Good if any(bc.get("uuid") == uuid for bc in beast_clients) else T.Disconnected
                 mlat_clients = a_dict.get("mlat_clients")
                 # print_err(f"alife returned {mlat_clients}")
                 if mlat_clients:
-                    self._mlat = (
-                        T.Yes
+                    api_mlat = (
+                        T.Good
                         if any(
                             (isinstance(mc.get("uuid"), list) and mc.get("uuid")[0] == uuid)
                             or (isinstance(mc.get("uuid"), str) and mc.get("uuid") == uuid)
                             for mc in mlat_clients
                         )
-                        else T.No
+                        else T.Disconnected
                     )
                 self._last_check = datetime.now()
             else:
@@ -295,11 +394,11 @@ class AggStatus:
                     tcp_uptime = int(result.stdout)
                 except:
                     print_err(f"WARNING: adsbx tcp uptime check returned: {result.stdout}")
-                self._beast = T.Yes if tcp_uptime > 5 else T.No
+                self._beast = T.Good if tcp_uptime > 5 else T.Disconnected
             except:
-                self._beast = T.No
+                self._beast = T.Disconnected
                 pass
-
+            """
             # now check mlat - which we can't really get easily from their status page
             # but can get from our docker logs again
             container_name = "ultrafeeder" if self._idx == 0 else f"uf_{self._idx}"
@@ -321,9 +420,10 @@ class AggStatus:
                 peer_count = make_int(match.group(1))
                 bad_sync_timeout = make_int(match.group(3))
                 # print_err(f"peer_count: {peer_count} bad_sync: {bad_sync_timeout}")
-                self._mlat = T.Yes if peer_count > 0 and bad_sync_timeout == 0 else T.No
+                self._mlat = T.Good if peer_count > 0 and bad_sync_timeout == 0 else T.Disconnected
             else:
                 self._mlat = T.Unknown
+            """
 
         elif self._agg == "tat":
             # get the data from the status text site
@@ -331,19 +431,21 @@ class AggStatus:
             tat_text, status = self.get_plain(text_url)
             if text_url and status == 200:
                 if re.search(r" No ADS-B feed", tat_text):
-                    self._beast = T.No
+                    self._beast = T.Disconnected
                 elif re.search(r"  ADS-B feed", tat_text):
-                    self._beast = T.Yes
+                    self._beast = T.Good
                 else:
                     print_err(f"can't parse beast part of tat response")
                     return
                 if re.search(r" No MLAT feed", tat_text):
-                    self._mlat = T.No
+                    api_mlat = T.Disconnected
                 elif re.search(r"  MLAT feed", tat_text):
-                    self._mlat = T.Yes
+                    api_mlat = T.Good
                 else:
                     print_err(f"can't parse mlat part of tat response")
+                    api_mlat = T.Unknown
                     # but since we got someting we could parse for beast above, let's keep going
+
                 self._last_check = datetime.now()
             else:
                 print_err(f"tat returned {status}")
@@ -352,7 +454,7 @@ class AggStatus:
             html_url = f"https://www.planespotters.net/feed/status/{uf_uuid}"
             ps_text, status = self.get_plain(html_url)
             if ps_text and status == 200:
-                self._beast = T.No if re.search("Feeder client not connected", ps_text) else T.Yes
+                self._beast = T.Disconnected if re.search("Feeder client not connected", ps_text) else T.Good
                 self._last_check = datetime.now()
             else:
                 print_err(f"planespotters returned {status}")
@@ -373,11 +475,15 @@ class AggStatus:
                 if not adsb or not mlat:
                     print_err(f"can't parse planewatch status {pw_dict}")
                     return
-                self._beast = T.Yes if adsb.get("connected") else T.No
-                self._mlat = T.Yes if mlat.get("connected") else T.No
+                self._beast = T.Good if adsb.get("connected") else T.Disconnected
+                api_mlat = T.Good if mlat.get("connected") else T.Disconnected
                 self._last_check = datetime.now()
             else:
                 print_err(f"planewatch returned {status}")
+        if api_mlat != self._mlat:
+            print_err(f"{self._agg}: api_mlat: {api_mlat} != self._mlat: {self._mlat}")
+            if self._mlat == T.Unknown:
+                self._mlat = api_mlat
 
     def __repr__(self):
         return f"Aggregator({self._agg} last_check: {str(self._last_check)}, beast: {self._beast} mlat: {self._mlat})"